changes to controllers

+++ controllers/article.js
@@ -15,6 +15,7 @@ module.exports = {
   // Create new article.
   create: async (req, res, next) => {
     try {
+      req.body.author = req.userId;
       const article = await Article.create(req.body);
       return res.json({ success: true, article });
     } catch (err) {
@@ -25,8 +26,13 @@ module.exports = {
   // Delete an existing article.
   delete: async (req, res, next) => {
     try {
-      const article = await Article.findByIdAndDelete(req.params.id);
-      return res.json({ success: true, article });
+      const art = await Article.findOne({_id: req.params.id});
+      if(art.author == req.userId) {
+        const article = await Article.findByIdAndDelete(req.params.id);
+        return res.json({ success: true, article });
+      } else {
+        return res.json({ msg: 'Unauthorized User' });
+      }
     } catch (err) {
       next(err);
     }
