+++ controllers/notes.js
@@ -5,6 +5,7 @@ module.exports = {
   // Create notes
   create: async (req, res, next) => {
     try {
+      req.body.author = req.userId;
       const note = await Note.create(req.body);
       return res.json({ success: true, note });
     } catch (err) {
@@ -34,13 +35,17 @@ module.exports = {
 
   // Update note
   update: async (req, res, next) => {
-    const {} = req.body;
     try {
-      const note = await Note.findByIdAndUpdate(
-        req.params.id,
-        req.body,
-        { new: true }
-      );
+      const nots = await Note.findOne({ _id: req.params.id });
+      if(nots.author == req.userId) {
+        const note = await Note.findByIdAndUpdate(
+          req.params.id,
+          req.body,
+          { new: true }
+        );
+      } else {
+        return res.json({ msg: 'Not Authorized' });
+      }
       return res.json({ success: true, note });
     } catch (err) {
       next(err);
@@ -50,8 +55,13 @@ module.exports = {
   // Delete note
   delete: async(req, res, next) => {
       try {
-        const note = await Note.findByIdAndDelete(req.params.id);
-        return res.json({ success: true, note });
+        const nots = await Note.findOne({ _id: req.params.id });
+        if(nots.author == req.userId) {
+          const note = await Note.findByIdAndDelete(req.params.id);
+          return res.json({ success: true, note });
+        } else {
+           return res.json({ msg: 'Not Authorized' });
+        }
       } catch(err) {
         next(err);
       }
